[test] wiring/no_fixture: move SPI tests into wiring/no_fixture_spi
